requirements.txt
Pillow>=10.3.0
pillow-avif-plugin>=1.4.3
